Extend Odyssey Quests
Patch with XML the ones that are easy.

Use a Transpiler on QuestNode_Root_Asteroid because it uses a readonly value to specify the default expiration, and quest construction is super weird.

Should clean up the settings menu and make the XML patch use the specified multiplier instead of hardcoding it to 2.