Add verboseMode setting, more cleanup
Add verbose mode setting and use it in MedicalDefaultsPatch.
Clean up method names, log messages, and inline some var declarations.
Reorder medical settings so that they match those in game. Even though I don't like the order the game has them, it's better to be consistent so as to not cause confusion when looking at the mod settings and then the medical defaults dialog.