Remove support for Quick Sleeper schedule modification
Not sure how this would work for user defined schedules: it sounds easy (just remove an hour or two of sleep time) but do you remove it at the start of the sleep schedule? the end? it's also very easy for a user to define a degenerate schedule (just put lots of separated sleep sections). If we remove sleep from every section, there would be little to no sleep. If we don't, what section should we remove? What if the user creates a triphasic schedule? Just too many edge cases.